add monitor and offline and realtime mode is ok.
