+++ HidProtocol/HIDTestDemo/HIDTestDemo.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
-VisualStudioVersion = 17.14.36401.2 d17.14
+VisualStudioVersion = 17.14.36401.2
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HIDTestDemo", "HIDTestDemo\HIDTestDemo.csproj", "{71AC80F0-CE7B-4C15-8F61-F7B9123C2794}"
 EndProject
